% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{NetworkWelfare}
\alias{NetworkWelfare}
\title{Estimate welfare effects}
\usage{
NetworkWelfare(
  policies,
  Y,
  X,
  Z,
  Z_int,
  D,
  W_0,
  W_0_train,
  cost = 0,
  focal_units = rep(1, length(Y)),
  model = "penalized_regression",
  family = "gaussian",
  params = list()
)
}
\arguments{
\item{policies}{vector of binary policies of length n}

\item{Y}{vector of outcomes. Y must contain missing values for those units whose neighbors are unknown but who are neighbors of focal points;}

\item{X}{variables for targeting;}

\item{Z}{variables used for estimating the conditional mean function and the propensity score;}

\item{Z_int}{variables included in the regression that interact with individual and neighbors' covariates;}

\item{D}{binary treatment assignments;}

\item{W_0}{adjacency matrix with 0/1 entries;}

\item{W_0_train}{adjacency matrix for training. Default is that W_0_train equals W_0.}

\item{cost}{cost of the treatment, default is zero;}

\item{focal_units}{vector of length n, indicating with 1 whether an individual is focal, and zero otherwise. An individual should be focal if all of her neighbors are observed. By default, all individuals are focal.}

\item{model}{model used for estimation. Any of 'penalized_regression', 'linear_regression', 'random_forest';}

\item{family}{family of the outcome, either gaussian or binary;}

\item{params}{list with additional parameters (see references)}
}
\value{
predictions: predicted effect on each individual;

welfare: point estimate of the welfare;
}
\description{
This function computes the welfare effect of a certain policy assignment.
}
